refactor navigationBar component
